Add promotion for books

+++ client/src/lib/useForm.js
@@ -12,7 +12,7 @@ export default function useForm(initial = {}) {
     function handleChange(e) {
         let { value, name, type } = e.target;
         if (type === 'number') {
-            value = parseInt(value);
+            value = parseFloat(value);
         }
         if (type === 'file') {
             [value] = e.target.files;
